fix: remake project for air-bnb requirements
