--- css/style.scss
@@ -1,66 +1,0 @@
-@import "reset";
-@import "normalize";
-
-.container {
-  margin: 0 auto;
-  width: 1000px;
-  .title {
-    text-align: center;
-    margin: 0;
-    padding: 20px 0;
-  }
-  .textarea {
-    margin: 0 auto;
-    padding: 5px;
-    width: 685px;
-    height: 150px;
-    border: solid 2px black;
-    border-radius: 5px;
-    resize: none;
-    display: block;
-    margin-bottom: 20px;
-  }
-  .keyboard {
-    margin: 0 auto;
-    width: 685px;
-    height: 235px;
-    border: solid 2px black;
-    border-radius: 5px;
-    display: flex;
-    flex-wrap: wrap;
-    padding: 5px;
-    gap: 5px;
-    .key {
-      padding-top: 10px;
-      width: 40px;
-      height: 40px;
-      text-align: center;
-      border: solid 2px black;
-      border-radius: 5px;
-      transition: 0.3s;
-      user-select: none; 
-      &:hover {
-        background: rgb(197, 197, 197);
-      }
-    }
-    .key.active {
-      border-radius: 20px;
-      background: rgb(197, 197, 197);
-    }
-    .Backspace,
-    .CapsLock,
-    .Enter,
-    .ShiftLeft,
-    .ShiftRight
-    {
-      width: 85px;
-    }
-    .Space {
-      width: 310px;
-    }
-  }
-  .description {
-    text-align: center;
-    margin: 20px 0;
-  }
-}
