+++ index.html
@@ -6,9 +6,9 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="shortcut icon" href="./assets/favicon.ico" type="image/x-icon">
-  <link rel="stylesheet" href="./css/style.css">
+  <link rel="stylesheet" href="./style.css">
   <title>Virtual Keyboard</title>
-  <script type="module" src="./js/script.js"></script>
+  <script type="module" src="./script.js"></script>
 </head>
 
 <body>
