--- js/keyboard-elements.js
@@ -1,84 +1,0 @@
-class KeyboardKeys {
-  constructor() {
-    this.Backquote = ['`', 'ё', '~', 'ё'];
-    this.Digit1 = ['1', '1', '!', '!'];
-    this.Digit2 = ['2', '2', '@', '"'];
-    this.Digit3 = ['3', '3', '#', '№'];
-    this.Digit4 = ['4', '4', '$', ';'];
-    this.Digit5 = ['5', '5', '%', '%'];
-    this.Digit6 = ['6', '6', '^', ':'];
-    this.Digit7 = ['7', '7', '&', '?'];
-    this.Digit8 = ['8', '8', '*', '*'];
-    this.Digit9 = ['9', '9', '(', '('];
-    this.Digit0 = ['0', '0', ')', ')'];
-    this.Minus = ['-', '-', '_', '_'];
-    this.Equal = ['=', '=', '+', '+'];
-    this.Backspace = ['Backspace', 'Backspace', 'Backspace', 'Backspace'];
-    this.Tab = ['Tab', 'Tab', 'Tab', 'Tab'];
-    this.KeyQ = ['q', 'й', 'q', 'й'];
-    this.KeyW = ['w', 'ц', 'w', 'ц'];
-    this.KeyE = ['e', 'у', 'e', 'у'];
-    this.KeyR = ['r', 'к', 'r', 'к'];
-    this.KeyT = ['t', 'е', 't', 'е'];
-    this.KeyY = ['y', 'н', 'y', 'н'];
-    this.KeyU = ['u', 'г', 'u', 'г'];
-    this.KeyI = ['i', 'ш', 'i', 'ш'];
-    this.KeyO = ['o', 'щ', 'o', 'щ'];
-    this.KeyP = ['p', 'з', 'p', 'з'];
-    this.BracketLeft = ['[', 'х', '[', 'х'];
-    this.BracketRight = [']', 'ъ', ']', ';'];
-    this.Backslash = ['\\', '\\', '|', '/'];
-    this.Delete = ['Delete', 'Delete', 'Delee', 'Delete'];
-    this.CapsLock = ['CapsLock', 'CapsLock', 'CapsLock', 'СapsLock'];
-    this.KeyA = ['a', 'ф', 'a', 'ф'];
-    this.KeyS = ['s', 'ы', 's', 'ы'];
-    this.KeyD = ['d', 'в', 'd', 'в'];
-    this.KeyF = ['f', 'а', 'f', 'а'];
-    this.KeyG = ['g', 'п', 'g', 'п'];
-    this.KeyH = ['h', 'р', 'h', 'р'];
-    this.KeyJ = ['j', 'о', 'j', 'о'];
-    this.KeyK = ['k', 'л', 'k', 'л'];
-    this.KeyL = ['l', 'д', 'l', 'д'];
-    this.Semicolon = [';', 'ж', ':', 'ж'];
-    this.Quote = ["'", 'э', '"', 'э'];
-    this.Enter = ['Enter', 'Enter', 'Enter', 'Enter'];
-    this.ShiftLeft = ['Shift', 'Shift', 'Shift', 'Shift'];
-    this.KeyZ = ['z', 'я', 'z', 'я'];
-    this.KeyX = ['x', 'ч', 'x', 'ч'];
-    this.KeyC = ['c', 'с', 'c', 'с'];
-    this.KeyV = ['v', 'м', 'v', 'м'];
-    this.KeyB = ['b', 'и', 'b', 'и'];
-    this.KeyN = ['n', 'т', 'n', 'т'];
-    this.KeyM = ['m', 'ь', 'm', 'ь'];
-    this.Comma = [',', 'б', '<', 'б'];
-    this.Period = ['.', 'ю', '>', 'ю'];
-    this.Slash = ['/', '.', '?', ','];
-    this.ArrowUp = ['&#8593;', '&#8593;', '&#8593;', '&#8593;'];
-    this.ShiftRight = ['Shift', 'Shift', 'Shift', 'Shift'];
-    this.ControlLeft = ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'];
-    this.MetaLeft = ['Win', 'Win', 'Win', 'Win'];
-    this.AltLeft = ['Alt', 'Alt', 'Alt', 'Alt'];
-    this.Space = [' ', ' ', ' ', ' '];
-    this.AltRight = ['Alt', 'Alt', 'Alt', 'Alt'];
-    this.ArrowLeft = ['&#8592;', '&#8592;', '&#8592;', '&#8592;'];
-    this.ArrowDown = ['&#8595;', '&#8595;', '&#8595;', '&#8595;'];
-    this.ArrowRight = ['&#8594;', '&#8594;', '&#8594;', '&#8594;'];
-    this.ControlRight = ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'];
-  }
-}
-
-const unhandleElements = [
-  'Tab',
-  'CapsLock',
-  'ShiftLeft',
-  'ControlLeft',
-  'MetaLeft',
-  'AltLeft',
-  'AltRight',
-  'ControlRight',
-  'ShiftRight',
-  'Enter',
-  'Delete',
-  'Backspace',
-];
-export { KeyboardKeys, unhandleElements };
