--- js/script.js
@@ -1,3 +1,0 @@
-import keyboardFunction from './keyboard.js';
-
-keyboardFunction();
